Fix DAB status printing
The strings for the DAB status dispplay are stored in PROGMEM. When using regular sprintf(), the string is not retrieved from the PROGMEM area. sprintf_P has a special formap specifier, %S, to be able to use strings in PROGMEM directly. Use that specifier to get status display working again.